refactor auth controller
