+++ lib/views/authView/AuthView.dart
@@ -4,13 +4,13 @@ import 'package:community_material_icon/community_material_icon.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'package:Videotheque/controllers/auth_controller/AuthController.dart';
+import 'package:Videotheque/controllers/AuthController/AuthController.dart';
 
 import '../app_view.dart';
 
 class AuthView extends StatelessWidget {
   AppView mainView;
-  
+
   AuthView(this.mainView);
 
   final OutlineInputBorder inputBorder = new OutlineInputBorder(
